tests/wpt/tests/css/css-position/parsing/left-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Positioned Layout Module Level 3: parsing left with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-position/#propdef-left">
<meta name="assert" content="left supports the full grammar 'auto | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("left", "auto");

test_valid_value("left", "-10px");
test_valid_value("left", "-20%");
test_valid_value("left", "calc(2em + 3ex)");
</script>
</body>
</html>
